This repository is for the PHP version
Play with these files and see if you can switch it over to PHP. I
haven't done much server-side stuff, so I think the only work will be
the includes for the topic menu. Oh, and the feedback tool in the menu,
but I can't even remember if that works right now because it isn't
important.